Split the tilesource source file into several files.
This should make it easier to maintain.
